refactor(utils): util 함수 분리
